Move parseEnumeratedAttribute to element.ts
