+++ packages/alfa-dom/src/node/element.ts
@@ -1,7 +1,13 @@
 import { Iterable } from "@siteimprove/alfa-iterable";
+import { Map } from "@siteimprove/alfa-map";
 import { Mapper } from "@siteimprove/alfa-mapper";
 import { None, Option, Some } from "@siteimprove/alfa-option";
+import {
+  EnumeratedValueError,
+  parseEnumeratedValue,
+} from "@siteimprove/alfa-parser";
 import { Predicate } from "@siteimprove/alfa-predicate";
+import { Err, Result } from "@siteimprove/alfa-result";
 import { Sequence } from "@siteimprove/alfa-sequence";
 
 import { Namespace } from "../namespace";
@@ -346,6 +352,32 @@ export namespace Element {
         Document.fromDocument(content)
       )
     );
+  }
+
+  /**
+   * Parse an enumerated attribute on an element (if it exists), according to a mapping.
+   * Mapping may includes special keys "missing" and "invalid"
+   * @see https://html.spec.whatwg.org/multipage/common-microsyntaxes.html#enumerated-attribute
+   */
+  export function parseEnumeratedAttribute<RESULT>(
+    name: string,
+    mapping: Map<string, RESULT>
+  ): (element: Element) => Option<RESULT> {
+    function parser(element: Element): Option<RESULT> {
+      return element
+        .attribute(name)
+        .map((attribute) => attribute.value)
+        .map(parseEnumeratedValue(mapping))
+        .getOr<Result<readonly [string, RESULT], EnumeratedValueError>>(
+          Err.of(EnumeratedValueError.Missing)
+        )
+        .mapOrElse(
+          ([_, result]) => Some.of(result),
+          (err) => mapping.get(err)
+        );
+    }
+
+    return parser;
   }
 }
